requirements.txt
# Core Training
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
datasets>=2.18.0
trl>=0.8.0

# Industry-Grade Optimization
# install unsloth via: pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
xformers>=0.0.26

# Utilities
pyyaml
python-dotenv
tqdm
packaging
psutil

# Logging & Experiment Tracking
tensorboard
wandb
scikit-learn
pandas
streamlit
fastapi
uvicorn
